+++ pkgs/firehose/lib/src/github.dart
@@ -34,14 +34,32 @@ class Github {
   String? get actor => _env['GITHUB_ACTOR'];
 
   /// Whether we're running withing the context of a GitHub action.
-  bool get inGithubContext => Platform.environment['GITHUB_ACTIONS'] != null;
+  bool get inGithubContext => _env['GITHUB_ACTIONS'] != null;
 
   /// The short ref name of the branch or tag that triggered the workflow run.
   /// This value matches the branch or tag name shown on GitHub. For example,
   /// `feature-branch-1`.
-  String? get refName => Platform.environment['GITHUB_REF_NAME'];
+  String? get refName => _env['GITHUB_REF_NAME'];
 
   http.Client get httpClient => _httpClient ??= http.Client();
+
+  /// Write the given [markdownSummary] content to the GitHub
+  /// `GITHUB_STEP_SUMMARY` file. This will cause the markdown output to be
+  /// appended to the GitHub job summary for the current PR.
+  ///
+  /// See also:
+  /// https://docs.github.com/en/actions/learn-github-actions/variables.
+  void appendStepSummary(String markdownSummary) {
+    var summaryPath = _env['GITHUB_STEP_SUMMARY'];
+    if (summaryPath == null) {
+      stderr.writeln("'GITHUB_STEP_SUMMARY' doesn't exist.");
+      return;
+    }
+
+    var file = File(summaryPath);
+    file.writeAsStringSync('${markdownSummary.trimRight()}\n\n',
+        mode: FileMode.append);
+  }
 
   Future<String> callRestApiGet(Uri uri) async {
     var token = githubAuthToken!;